MoveOperation bugfix for CLI overrides (#36)
## Context

As pointed out in https://github.com/gtronset/beets-filetote/issues/35,
certain configuration scenarios can result in the `MoveOperation`
selection is `None`. In that issue, @seanbreckenridge is right that this
comes down to _when_ the plugin defines (really, captures) the desired
Operation (copy, move, etc.) that is occurring in the import.

Previously, this was captured on initialization in the constructor for
Filetote. However, due to the lifecycle of `beets`, this is called
before the `import` command initializes/runs. As a side effect, any of
the overriding CLI options passed along (relevant here, `--move` or
`--copy`) would set/override the operation on the media file(s) but,
since the operation was defined before this and not updated, the extra
files that Filetote would manage would utilize a different operation
than the media file.

In addition, if the `config.yml` explicitly did _not_ set an operation
(meaning, all were `false`) and if one were to rely on the CLI only, the
`MoveOperation` would result in `None` (the plugin initializes with
`None` which matches the `config.yml` but the operation is not updated
once the import starts to grab the option provided in the CLI).

## Other notes

`beets` prioritize the operation as the options are mutually exclusive.
There are three layers to how this occurs:

* First, it looks at values within the `config.yml`, prioritizing in
this order
([`importer.py`](https://github.com/beetbox/beets/blob/master/beets/importer.py#L1572-L1583)):
  * `move`
  * `copy`
  * `link`
  * `hardlink`
  * `reflink`
* If a CLI option is provided, it uses that (either `--move` or
`--copy`):
* If both CLI options (`--move` and `--copy`) are provided, it
prioritizes `copy`
([`ui/commands.py`](https://github.com/beetbox/beets/blob/master/beets/ui/commands.py#L983-L986))

There are a few more nuances, but this covers the vast majority of
cases.

## Changes

This moves the setting of the operation used by Filetote to the
`import_begin` event that is triggered at the start of the import
process. This also updates/adds tests for operations provided via CLI.

This also includes the suggested change from
https://github.com/gtronset/beets-filetote/pull/34 which is, indeed,
helpful for debugging.

---

Big thanks to @seanbreckenridge for running much of this down and
documenting their journey in such detail!