Bugfix of mediafile types - Exclude m4a, m4b, etc. (#43)
Filetote, like `beets-copyartifacts` and `beets-extrafiles` before it,
relied on
[MediaFile.TYPES](https://github.com/beetbox/mediafile/blob/master/mediafile.py#L68-L82)
for determining which files are unimported music/media files vs.
non-music extra files and artifacts. However, that list isn't intended
to describe the _extensions_ of music files, which meant there are
certain extensions (specifically,`.m4a`, `.wma`, and `.wave`) that are
processed by beets but not otherwise considered a music file by that
dictionary alone. This augments the list to include the missing
extensions so that they are not also grabbed by Filetote in its
processing.

This also augments/includes `.m4b` for those using the excellent
[beets-audible](https://github.com/Neurrone/beets-audible) plugin, which
allows for `.m4b` support in beets for audiobooks.