Misc. Refactors to Filetote (#51)
This is a reasonably substantial reworking and refactoring of the
plugin, primarily focused on better organization and utilizing more core
beets functionality for "mapping" properties for the extra
files/artifacts. This improves typing in the plugin quite a bit and
introduces dataclasses for better organization of certain data. Lastly,
this flattens some functions and removes several linting overrides.

A minor bug relating to path encoding was also fixed.