Allow paired files to be by ext (#54)
This adds the ability to be specific about which extensions should
qualify as being a "pair", allowing for targeted filetypes instead of
having to rely upon the global Filetote extension list.

These extensions are additive and independent from the global list,
meaning those specified in the `pairing` config node will apply even if
they do not otherwise appear in the global list. This can be useful for
only targeting paired files, or only caring about certain file types if
they are pairs.

This introduces a breaking change to the config of pairs. `pairing` has
been converted from a boolean to an object with other like-config. Take
the following config:

```yaml
filetote:
  pairing: true
  pairing_only: false
```

These will both now be represented as individual settings within
`pairing`:

```yaml
filetote:
  pairing:
    enabled: true
    pairing_only: false
    extensions: ".lrc"
```

Both remain optional and both default to `false`.