Fix "cannot find the file specified" bug (#115)
This addresses the following warnings and a bug in the hidden field of
`filesize` for Items. The warnings would manifest with the following
types of logs (macOS and Windows):

> could not get filesize: [Errno 2] No such file or directory:

> could not get filesize: [WinError 2] The system cannot find the file
specified:

This PR should fix (some) problems mentioned in these issues:
* https://github.com/gtronset/beets-filetote/issues/99
* https://github.com/gtronset/beets-filetote/issues/111

This is caused by a timing issue from when mapped data for the extra
files and artifacts fetching the available metadata from the source
Item, which occurs when beets tries to grab data of filesize here:
https://github.com/beetbox/beets/blob/708a04d4a88eb955f261a21f0b1146c9e2ae0a55/beets/library.py#L949-L958

The functions in Filetote that give access to all the [available
values](https://beets.readthedocs.io/en/stable/reference/pathformat.html#available-values)
of the media item combined with when Filetote moves items (after the
media item has already moved) leads to a side-effect of Beets trying to
get the filesize metadata from the old media item's file location. I've
updated Filetote to grab the metadata _before_ the item is moved to
resolve this (using the `before_item_moved` event). From what I can
tell, this issue only exists on `move`, as the original file location
still exists in all other operations.

Interestingly, there is [an additional "hidden" attribute value of
`filesize`](https://github.com/beetbox/beets/blob/708a04d4a88eb955f261a21f0b1146c9e2ae0a55/beets/library.py#L666)
that before this fix would result in `0` for anyone was using
`$filesize` and the `move` operation. This should now work as expected
for all operations.

A test was added to explicitly check for this error and to ensure the
value returned in non-zero.