Reevaluate & fix types (now with stubs!) (#119)
## Overview

This is a comprehensive reevaluation of the type hints used by Filetote.
With recent excellent efforts by Beets contributors (ex:
https://github.com/beetbox/beets/pull/4591), much more of beets has
typing which means less guessing needs to occur and that stubs can be
built out (recognizing the type hints added to beets are not yet
released; stubs for it will be removed at a later date once possible).

This adds Type Hint Stubs for the following, which allows for the
removal of `--ignore-missing-imports` from MyPy:
* [beets](https://github.com/beetbox/)
* [confuse](https://github.com/beetbox/confuse/)
* [mediafile](https://github.com/beetbox/mediafile)
*
[reflink](https://gitlab.com/rubdos/pyreflink/-/tree/master?ref_type=heads)

This also adds linting for `.pyi` (stub) files courtesy of
[flake8-pyi](https://pypi.org/project/flake8-pyi). As part of that,
Flake8's config was extended for better compatibility with other
linting/formatting (see:
https://github.com/gtronset/beets-filetote/pull/119/commits/e5f94b3f27a904d79b7a628975c2989c605df675).

This also adds [typeguard](https://github.com/agronholm/typeguard) to
run during pytests. Note: that functionality is disabled in Python 3.6
due to a bug in the last version of `typeguard` for it (see:
https://github.com/agronholm/typeguard/issues/168).

Finally, validation for the Filetote config has been added to ensure
that at least the correct types are being set in the config itself.

Bonuses: deprecation warning for libraries used by `confuse` and
`mediafile` are now ignored in Pytest; CI's running of pytest will be
slightly faster.

---

## A note on Paths

[Beets has a long blog post](https://beets.io/blog/paths.html) on its
strategy of handling path types; TL;DR they've opted to handle them
everywhere as bytestrings (`bytes`). This plugin (and its predecessor
`copyartifacts`) would use them interchangeably. This has been updated
with the help of better type hints and the changes mentioned above.

[While not documented
explicitly](https://beets.readthedocs.io/en/v1.6.0/dev/plugins.html#listen-for-events),
`MoveOperation`-related events expect `bytes`-type paths for their
`source` and `destination` parameters. This was discovered by digging
deeper here:
https://github.com/beetbox/beets/blob/708a04d4a88eb955f261a21f0b1146c9e2ae0a55/beets/library.py#L915

---

## Bugs fixed in this:

* https://github.com/gtronset/beets-filetote/issues/111 (effectively the
same issue as mentioned here:
https://github.com/adammillerio/beets-copyartifacts/pull/6), caused by
improper type handling
* Improper translation of the `pairing` setting/node
(`filetote.pairing.extensions` would not come through as a string
sequence)

---

## Future Changes

Additional refactoring is needed in the core handling of files to
simplify and make it easier to understand. In particular relevance to
this PR, there might be additional typing issues hidden in some of thee
deeper parts that were not addressed here, due to the nature of how
those methods are written.