Enable Filetote on "move" command (#124)
This change originated as a fix for
https://github.com/gtronset/beets-filetote/issues/110. As it turns out,
Filetote was limited to operations just during _import_ (i.e., the
`import` CLI command). This updates the functionality to other commands
that change media files' location in the file system (move, copy, etc.),
such as the `move` or `modify` commands.

This behavior is a remnant from `copyartifacts` which effectively
oriented the plugin to the _import_ process in particular. However, this
is understandably confusing as other CLI operations do similar actions
as import and can very much impact the Beets Library and underlying
media files.

---

Note: pruning after the fact may not clean up the folder if empty. This
is likely due to it still relying on `import_path` as a concept, which
is `None` for these other CLI commands. Address this (along with other
reports of similar outcomes) will occur in a follow-up PR.