Fix reimport via query (IndexError bugfix) (#126)
There was a report of crashes when re-importing a library
(https://github.com/gtronset/beets-filetote/issues/82). While tests did
exist for reimports, specific tests for reimporting via a query did not.
This issue arose from an erroneous assumption that all imports would
always have an import path specified, which isn't the case when a query
is provided.

This adds a guard during imports to ensure path(s) exist before trying
to grab the first value, along with a test.

---

Similar to other recent PRs, it appears that pruning after moves/copies
are complete likely do not clean up the folder when a query is used.
This does appear to be due to pruning relying on `import_path` as a
concept, which is `None` in these cases. Addressing this will occur in a
follow-up PR.