Bugfix for Update CLI command (#133)
There was a request to ensure that the [`update`
command](https://beets.readthedocs.io/en/v1.6.0/reference/cli.html#update)
works when files are moved (see
https://github.com/gtronset/beets-filetote/pull/124#issuecomment-1840293273).
This PR adds tests to ensure that does occur.

By default, media files are automatically renamed based on their new
metadata. The work done in
https://github.com/gtronset/beets-filetote/pull/124 enables any CLI
command that results in moving files to also enable Filetote to move
artifacts and extra files. However, it looks like `update` will perform
two moves, first applying the update to the beets `Item` and second to
the `Album` ([at least in beets
`1.6.0`](https://github.com/beetbox/beets/blob/v1.6.0/beets/ui/commands.py#L1161-L1198)).
If there's a change to the artist, the first move uses the `comp` Path
settings, but the second back to the default (the correct final
location). The bug occurs since Filetote doesn't expect multiple moves
to occur, and the previous configuration resulted in the artifacts being
moved to the first location but not the second (a side effect of
handling file operations on CLI exit, which ideally will be changed to
incremental soon).

This updates Filetote to track the Item more in-depth and, if it moves
more than once, updates the destination of the artifact.