Test and ensure subdirectory with $albumpath (#141)
<!-- markdownlint-disable-file MD041 -->
## Description

This tests and ensures that there's flexibility in moving extra files to
different needs, such as different directories above the music file it's
relocated with. This adds documentation to better explain this, which
should help address such needs as explained in
https://github.com/Holzhaus/beets-extrafiles/issues/18.

As part of this work, the `inline` plugin was explored to help provide a
universal method of moving to the parent directory. However, since
`$ablumpath` is computed as part of Filetote's run and since it looks
like the output from `inline` is sanitized, I was unable to discover a
path that worked reliably there. That said, this might still be viable
in the future.

For now, the implication is that for all OS's there's a way to be as
specific as needed, which have been documented as part of this PR.

## To Do

<!--
- If you believe one of below checkpoints is not required for the change
you
are submitting, cross it out and check the box nonetheless to let us
know.
  For example: - [x] ~Changelog~
- Regarding the changelog, often it makes sense to add your entry only
after code
review is finished. That way you might prevent conflicts from other PR's
in
that file, as well as keep the chance high your description fits with
the
  latest revision of your feature/fix.
- Regarding documentation, bugfixes often don't require additions to the
docs.
-->

- [x] Documentation (update `README.md`)
- [x] Changelog (add an entry to `CHANGELOG.md`)
- [x] Tests