Allow for retaining path hierarchy (#144)
<!-- markdownlint-disable-file MD041 -->
## Description

This PR allows for the retention of artifact subfolder path hierarchy.
For instances where there can be unpredictable and potentially multiple
layers of subfolders within a given album path, those same subfolders
can now be used after the files are manipulated/moved. Before this PR,
items could be correctly grabbed from the source folder but would
collect at either the album directory (effectively flattened) or under a
specific path specification if defined as part of a pattern or globally
for all of a given extension. Since different files might live in
different locations, this might lead to conflicts and certain files
being moved and is lacking in the ability to customize.

This originated as a feature request in
https://github.com/gtronset/beets-filetote/discussions/143.

## To Do

<!--
- If you believe one of below checkpoints is not required for the change
you
are submitting, cross it out and check the box nonetheless to let us
know.
  For example: - [x] ~Changelog~
- Regarding the changelog, often it makes sense to add your entry only
after code
review is finished. That way you might prevent conflicts from other PR's
in
that file, as well as keep the chance high your description fits with
the
  latest revision of your feature/fix.
- Regarding documentation, bugfixes often don't require additions to the
docs.
-->

- [x] Documentation (update `README.md`)
- [x] Changelog (add an entry to `CHANGELOG.md`)
- [x] Tests