Remove need for typing_extensions in plugin (#146)
<!-- markdownlint-disable-file MD041 -->
## Description

This removes the need for the
[typing_extensions](https://pypi.org/project/typing-extensions/) to be
installed on older versions of Python (previously required for versions
<3.8).

`typing_extensions` was needed to support the `Literal` and `TypeAlias`
type annotations. While `TypeAlias` was a convenience, `Literal` helped
to restrict certain default or special string values for the Filetote
configuration. This was replaced with the more generic `str` to reduce
this module dependency. Once Python 3.6 and 3.7 are dropped by beets
(next beets release, TBD when), `Literal` can be used again which will
re-allow stricter and more accurate typing.

This PR resolves the issue
https://github.com/gtronset/beets-filetote/issues/145.

## To Do

<!--
- If you believe one of below checkpoints is not required for the change
you
are submitting, cross it out and check the box nonetheless to let us
know.
  For example: - [x] ~Changelog~
- Regarding the changelog, often it makes sense to add your entry only
after code
review is finished. That way you might prevent conflicts from other PR's
in
that file, as well as keep the chance high your description fits with
the
  latest revision of your feature/fix.
- Regarding documentation, bugfixes often don't require additions to the
docs.
-->

- [x] Documentation (update `README.md`)
- [x] Changelog (add an entry to `CHANGELOG.md`)
- [x] ~~Tests~~