Update beets version & other dependency requirements (#155)
<!-- markdownlint-disable-file MD041 -->
## Description

The specification for Beets version was previously not allowing for the
installation beyond the current released version of `1.6.0`. This has
been updated, along with other dependencies in the `pyproject.toml`.

This resolves https://github.com/gtronset/beets-filetote/issues/153.

## To Do

<!--
- If you believe one of below checkpoints is not required for the change
you
are submitting, cross it out and check the box nonetheless to let us
know.
  For example: - [x] ~Changelog~
- Regarding the changelog, often it makes sense to add your entry only
after code
review is finished. That way you might prevent conflicts from other PR's
in
that file, as well as keep the chance high your description fits with
the
  latest revision of your feature/fix.
- Regarding documentation, bugfixes often don't require additions to the
docs.
-->

- [X] Documentation (update `README.md`)
- [X] Changelog (add an entry to `CHANGELOG.md`)
- [X] ~~Tests~~