Add tests to ensure priority and order for `paths` queries (#185)
## Description

This investigates issues described in
https://github.com/gtronset/beets-filetote/issues/159. Based on
additional tests and replication testing, the issue described there is
likely no longer present (best guess that it was either: (a)
previously/already fixed; (b) due to Python 3.6 not respecting order in
dictionaries). That said, this PR adds explicit testing and additional
documentation around `path` renaming preferential order. This also
slightly refactors the `paths` to ensure the processing order examines
those queries provided in the Filetote `paths` first (though this should
be irrelevant, since the only time this would matter in within patterns,
but the pattern is predetermined based on the `patterns` config order
when matching occurs).

## To Do

- [x] Documentation (update `README.md`)
- [x] Changelog (add an entry to `CHANGELOG.md`)
- [X] Tests