Add better Workflow caching & other CI performance tweaks (#191)
## Description

Changes in https://github.com/gtronset/beets-filetote/pull/188 led to
the inclusion of FFmpeg in the test suite. Unfortunately, the
installation time of FFmpeg increased the average test run from ~2.5
minutes to 3.5-4+ minutes. This PR adds in better caching, removal of
Tox being the environment to run tests (instead use Poetry to run Pytest
and linting-related tasks directly), and a few other tweaks. The result
is much more efficient run times for the testing and linting jobs (tests
are down to ~1-1.5 minutes; listing has decreased ~30% from ~30s to
~20).

### Main Changes

The largest change is directly installing the executable for FFmpeg,
recognizing potential slowness from installing via either APT or
Chocolatey (especially!). Instead, binaries are installed from
recommended sources to a catchable location, thus allowing significantly
quicker steps in the workflow run. The version of FFMpeg is hardcoded to
the most recent universally available version, meaning it will need to
be updated over time.

This relies on Poetry installations via Pipx, which is then cached,
applying not just to the "CI" workflow but also the "Release" one. An
exploration of Tox caching was done along with checking the timing of
Tox parallelism, but ultimately, it was deemed less performant compared
to Poetry. For future reference, caching can be achieved with code such
as that removed in
https://github.com/gtronset/beets-filetote/pull/191/commits/978f3d984ae7ed81194b3dcf65fc32d299fb3423.

`Act` configuration has also had a few tweaks to improve ergonomics,
including caching capabilities for macOS and easier secret/variable
specification via the now-included `.actrc`.

---

### Notes

This type of work requires quite a bit of trial and error due to the
nature of GitHub Actions, different runners, etc., and thus many
exploration commits to trigger CI.

Caching of FFmpeg has minor differences between Linux and Windows. While
this code could be DRY-er, I've opted to keep the separate for now to
ease in maintenance.

## To Do

- [X] ~Documentation (update `README.md`)~
- [x] Changelog (add an entry to `CHANGELOG.md`)
- [X] ~Tests~